refactor: TaxInvoice 도메인 중심으로 병합 #110
