+++ src/main/java/Seoul_Milk/sm_server/domain/taxInvoice/enums/ThreadTerm.java
@@ -1,4 +1,4 @@
-package Seoul_Milk.sm_server.domain.taxInvoiceValidation.enums;
+package Seoul_Milk.sm_server.domain.taxInvoice.enums;
 
 public enum ThreadTerm {
     THREAD_TERM(1000L);
